+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ###
 - Support for optionflags in .. testoutput::
+- Optionflags from pytest.ini are used as the default options
 
 ## [0.1.2] - 2017-11-04
 ### Added